application-one/requirements.txt
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_sqs
aws-cdk.aws_sns
aws-cdk.aws_sns_subscriptions
aws-cdk.aws_ecs_patterns
aws-cdk.aws_ecr
aws-cdk.aws_s3